More progress on inputs and combiner, created ecd module and moved inputs and outputs functions there
