+++ ludwig/features/bag_feature.py
@@ -168,4 +168,4 @@ class BagInputFeature(BagBaseFeature, InputFeature):
 
     @staticmethod
     def populate_defaults(input_feature):
-        set_default_value(input_feature, 'tied_weights', None)
+        set_default_value(input_feature, TIED, None)
