+++ ludwig/features/binary_feature.py
@@ -114,7 +114,7 @@ class BinaryInputFeature(BinaryBaseFeature, InputFeature):
 
     @staticmethod
     def populate_defaults(input_feature):
-        set_default_value(input_feature, 'tied_weights', None)
+        set_default_value(input_feature, TIED, None)
 
 
 class BinaryOutputFeature(BinaryBaseFeature, OutputFeature):
