+++ ludwig/features/category_feature.py
@@ -173,7 +173,7 @@ class CategoryInputFeature(CategoryBaseFeature, InputFeature):
 
     @staticmethod
     def populate_defaults(input_feature):
-        set_default_value(input_feature, 'tied_weights', None)
+        set_default_value(input_feature, TIED, None)
 
 
 class CategoryOutputFeature(CategoryBaseFeature, OutputFeature):
