+++ ludwig/features/date_feature.py
@@ -177,7 +177,7 @@ class DateInputFeature(DateBaseFeature, InputFeature):
 
     @staticmethod
     def populate_defaults(input_feature):
-        set_default_value(input_feature, 'tied_weights', None)
+        set_default_value(input_feature, TIED, None)
 
 
 date_encoder_registry = {
