+++ ludwig/features/h3_feature.py
@@ -139,7 +139,7 @@ class H3InputFeature(H3BaseFeature, InputFeature):
 
     @staticmethod
     def populate_defaults(input_feature):
-        set_default_value(input_feature, 'tied_weights', None)
+        set_default_value(input_feature, TIED, None)
 
 
 h3_encoder_registry = {
