+++ ludwig/features/image_feature.py
@@ -414,7 +414,7 @@ class ImageInputFeature(ImageBaseFeature, InputFeature):
 
     @staticmethod
     def populate_defaults(input_feature):
-        set_default_value(input_feature, 'tied_weights', None)
+        set_default_value(input_feature, TIED, None)
         set_default_value(input_feature, 'preprocessing', {})
