+++ ludwig/features/sequence_feature.py
@@ -204,7 +204,7 @@ class SequenceInputFeature(SequenceBaseFeature, InputFeature):
 
     @staticmethod
     def populate_defaults(input_feature):
-        set_default_value(input_feature, 'tied_weights', None)
+        set_default_value(input_feature, TIED, None)
         set_default_value(input_feature, 'encoder', 'parallel_cnn')
