+++ ludwig/features/set_feature.py
@@ -176,7 +176,7 @@ class SetInputFeature(SetBaseFeature, InputFeature):
 
     @staticmethod
     def populate_defaults(input_feature):
-        set_default_value(input_feature, 'tied_weights', None)
+        set_default_value(input_feature, TIED, None)
 
 
 class SetOutputFeature(SetBaseFeature, OutputFeature):
