+++ ludwig/features/timeseries_feature.py
@@ -181,7 +181,7 @@ class TimeseriesInputFeature(TimeseriesBaseFeature, SequenceInputFeature):
 
     @staticmethod
     def populate_defaults(input_feature):
-        set_default_value(input_feature, 'tied_weights', None)
+        set_default_value(input_feature, TIED, None)
 
 
 # this is still WIP
