+++ ludwig/features/vector_feature.py
@@ -166,7 +166,7 @@ class VectorInputFeature(VectorBaseFeature, InputFeature):
 
     @staticmethod
     def populate_defaults(input_feature):
-        set_default_value(input_feature, 'tied_weights', None)
+        set_default_value(input_feature, TIED, None)
         set_default_value(input_feature, 'preprocessing', {})
